added a times in view

+++ db/migrate/20110727175427_add_time_ran_to_configurations.rb
@@ -1,6 +1,6 @@
 class AddTimeRanToConfigurations < ActiveRecord::Migration
   def self.up
-    add_column :configurations, :time_ran, :float
+    add_column :configurations, :time_ran, :integer
   end
 
   def self.down
